# public/api_doc/swagger.json
{
    "openapi": "3.0.0",
    "paths": {
...
        "/api/empleado/{id}": {
            "get": {
                "tags": [
                    "Empleado"
                ],
                "summary": "Recuperación de un empleado por id",
                "operationId": "74b741860b316c27f712e6fbf4918774",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Identificador del empleado que se quiere recoger. Rango posible de valores: números enteros desde el 1 hasta el +∞",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "404": {
                        "description": "El empleado no existe",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 404
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Empleado no encontrado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "El empleado ha sido devuelto correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                },
                                "example": {
                                    "id": "3",
                                    "nombre": "string",
                                    "apellidos": "string",
                                    "correo": "string"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Empleado"
                ],
                "summary": "Edición de un empleado",
                "operationId": "2c652af74a4eb346da2646e4a79fc3d1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Identificador del empleado que se quiere editar. Rango posible de valores: números enteros desde el 1 hasta el +∞",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Información a editar del empleado",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "string"
                                    },
                                    "apellidos": {
                                        "type": "string",
                                        "example": "string"
                                    },
                                    "correo": {
                                        "type": "string",
                                        "example": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "404": {
                        "description": "El empleado no existe",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 404
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Empleado no encontrado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "El empleado ha sido editado correctamente. Se devuelven los nuevos datos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 3
                                            },
                                            "nombre": {
                                                "type": "string",
                                                "example": "string"
                                            },
                                            "apellidos": {
                                                "type": "string",
                                                "example": "string"
                                            },
                                            "correo": {
                                                "type": "string",
                                                "example": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "example": {
                                    "message": "El empleado ha sido editado correctamente",
                                    "data": {
                                        "id": "3",
                                        "nombre": "string",
                                        "apellidos": "string",
                                        "correo": "string"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
    }
}